% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{available}
\alias{available}
\title{Look-up available programs for installation}
\usage{
available(pattern = NULL)
}
\arguments{
\item{pattern}{REGEX to limit returns}
}
\value{
data.frame
}
\description{
Returns a data.frame listing all available external programs
that can be installed with \code{containr}.
}
\details{
Limit the names of programs returned using the \code{pattern}
argument.
}
\seealso{
Other user: \code{\link{details}}, \code{\link{remove}}
}
