% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContainR_package.R
\docType{package}
\name{ContainR}
\alias{ContainR}
\alias{ContainR-package}
\title{aggreCAT: mathematically aggregating expert judgments}
\description{
To learn more about \pkg{ContainR}, start with the vignettes: \code{browseVignettes(package = "ContainR")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/psychtek/containR}
  \item Report bugs at \url{https://github.com/psychtek/containR/issues}
}

}
\author{
\strong{Maintainer}: Aaron Willcox \email{aaron@willcox.io} (\href{https://orcid.org/0000-0003-2536-2596}{ORCID})

}
\keyword{internal}
