% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{create_config_file}
\alias{create_config_file}
\title{Create Config File}
\usage{
create_config_file(project_name)
}
\arguments{
\item{project_name}{Given name of project. Defaults to repo directory.}
}
\description{
Creates a temp .Rprofile and writes a setHook to launch
activate current project in the Rstudio docker container.
}
