% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_rocker_table}
\alias{data_rocker_table}
\title{data_rocker_table}
\format{
A data frame:
\describe{
\item{name}{name of rocker image}
\item{image}{image of rokcer stack}
\item{base_image}{what the image is build on}
\item{description}{Short description}
}
}
\source{
\url{https://rocker-project.org/images/}
}
\usage{
data_rocker_table
}
\description{
A data table of rocker versioned images
}
\keyword{datasets}
