% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{details}
\alias{details}
\alias{details}
\alias{install}
\title{Look-up details on program}
\usage{
details(id)

install(id)
}
\arguments{
\item{id}{Program ID}

\item{id}{Program ID}
}
\description{
Prints to console detailed information on program available for
installation such as build status, version number, maintainer and program
notes.

Prints to console detailed information on program available for
installation such as build status, version number, maintainer and program
notes.
}
\seealso{
Other user: \code{\link{available}}, \code{\link{remove}}

Other user: \code{\link{available}}, \code{\link{remove}}
}
