% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker}
\alias{docker}
\title{docker Class Object}
\description{
A R6 Docker class as a wrapper for the docker commands.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-docker-new}{\code{docker$new()}}
\item \href{#method-docker-print}{\code{docker$print()}}
\item \href{#method-docker-show_output}{\code{docker$show_output()}}
\item \href{#method-docker-show_json}{\code{docker$show_json()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docker-new"></a>}}
\if{latex}{\out{\hypertarget{method-docker-new}{}}}
\subsection{Method \code{new()}}{
Work in progress: Eventually will be part of the containr class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$new(process = NA, commands = NULL, options = NULL, args = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process}}{The system process. Plan is to add \code{Podman} support but currently only
setup for \code{docker}. Although Podman commands are formatted in a similar context.}

\item{\code{commands}}{The main docker commands. Currently support commands
are "image", "container", "search", "build", "history", "inspect".}

\item{\code{options}}{the \code{command} option arguments or \code{image}}

\item{\code{args}}{Additional arguments or flags for the docker commands}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docker-print"></a>}}
\if{latex}{\out{\hypertarget{method-docker-print}{}}}
\subsection{Method \code{print()}}{
Prints the full command to the console for viewing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docker-show_output"></a>}}
\if{latex}{\out{\hypertarget{method-docker-show_output}{}}}
\subsection{Method \code{show_output()}}{
Displays returned output in \code{tibble} format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$show_output()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docker-show_json"></a>}}
\if{latex}{\out{\hypertarget{method-docker-show_json}{}}}
\subsection{Method \code{show_json()}}{
Show the json in pretty format from the command
Formatting Function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$show_json()}\if{html}{\out{</div>}}
}

}
}
