% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_build}
\alias{docker_build}
\title{Docker build}
\usage{
docker_build(dockerfile = NULL, name = NULL)
}
\arguments{
\item{dockerfile}{Character path to supplied Dockerfile. Save to \code{inst/dockerfiles}
if one doesnt exists then a Dockerfile can be generated from the \code{docker_build} function.}

\item{name}{Name and or name and tag of container name such as username/image_name}
}
\description{
Builds a docker image from a file location
}
