% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_check}
\alias{docker_check}
\title{Check Docker Services}
\usage{
docker_check()
}
\description{
Performs a quick system check for installation
of Docker and if not found, gives URL to page for install instructions.
}
