% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_create}
\alias{docker_create}
\title{Create a Dockerfile}
\usage{
docker_create(
  dockerfile = "inst/dockerfiles/Dockerfile",
  which_pkgs = c("loaded", "installed"),
  rocker_name = "rstudio",
  tag = NULL,
  include_python = TRUE
)
}
\arguments{
\item{dockerfile}{Default location is in the \verb{inst/dockerfiles/} folder. Build will save the final \code{Dockerfile}
to this location. You can also add youre own Dockerfile, however, this package was primarily designed to launch
an active project into a Rstudio container from rocker version.}

\item{which_pkgs}{Provide a string of either \code{loaded} or \code{installed}. \code{Loaded} will included packages that are currently
loaded in your active project session. The \code{installed} string will included everything inside you local default R library.}

\item{rocker_name}{The Rocker name of the stack that you want to build off. To see the current stack options
load the \link{data_rocker_table}. A character string from the \code{name} will load the \code{FROM} command with the associated \code{image}.}

\item{tag}{A character string of the require version. If no tag is supplied then the function will get the local
version of R from \code{getRversion()},}

\item{include_python}{Flag to install python using the rocker scripts \url{https://github.com/rocker-org/rocker-versioned2} which have had minor modifications. Future updates
will see this streamlined.}
}
\description{
This function creates a dockerfile based off the rocker version image stacks and reads
from session information to included loaded or installed R libraries.
}
