% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_deploy}
\alias{docker_deploy}
\title{Title}
\usage{
docker_deploy(image, system_deps = c(), working_dir = getwd(),
  packrat_install_method = c("cran", "github"))
}
\arguments{
\item{working_dir}{}
}
\description{
Title
}
