% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{docker_folder_setup}
\alias{docker_folder_setup}
\title{Setup Docker Files}
\usage{
docker_folder_setup(print_tree = FALSE)
}
\arguments{
\item{print_tree}{Flag to display the directory tree}
}
\description{
Setups up a dockerfile and scripts folders in root directory
of the active R project or directory.
}
