% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_login}
\alias{docker_login}
\title{Docker Login}
\usage{
docker_login(username)
}
\arguments{
\item{username}{Character string of dockerhub user login}
}
\description{
Log into docker hub using a \code{DOCKER_PAT} that should be setup
through \url{https://hub.docker.com} and assigned to \code{.Renviron}.
}
