% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_pull}
\alias{docker_pull}
\title{Pull Rocker Image}
\usage{
docker_pull(rocker_verse, tag = NULL)
}
\arguments{
\item{rocker_verse}{Rocker image name from the \link{data_rocker_table}}

\item{tag}{Version tag of rocker image. Will default to local
r version using getRversion.}
}
\description{
Pull Rocker Image
}
