% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R, R/run_docker.R
\name{docker_run}
\alias{docker_run}
\title{RStudio Addin: Containerize and Render R Markdown Document with Docker}
\usage{
docker_run(image = "rstudio", tag = NULL)

docker_run(image = "rstudio", tag = NULL)
}
\arguments{
\item{image}{Name of rocker image from \link{data_rocker_table}. Defaults to
the \code{rstudio} stack \code{rocker/rstudio:latest}. For more information about each stack
visit \url{https://rocker-project.org/images/}. If you ran \code{docker_build()} then supply
this argument with the string of the name of the build.}

\item{tag}{Version number of stack. Defaults to the \verb{:latest} tag if none is supplied.}
}
\description{
RStudio addin for containerizing and rendering the current document.

Launches current active R project into a rocker container
and opens a browser.
}
\keyword{internal}
