% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_run_custom}
\alias{docker_run_custom}
\title{Run Replicats RStudio Session}
\usage{
docker_run_custom(docker_image = NULL)
}
\arguments{
\item{docker_image}{A docker container image to launch current project into.
This defaults to the \code{psychtek/replicats} which includes rstudio, python and
additional libraries and dependencies.}
}
\description{
\code{docker_run_custom} runs the current active project in a rstudio
container. Present it defaults to the \code{replicats-docker} image. The session
will also copy the local user .Renviron and rstudio config directory.
}
