% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_search}
\alias{docker_search}
\title{Docker Search}
\usage{
docker_search(search_string = "rstudio")
}
\arguments{
\item{search_string}{Character string for search term}
}
\description{
Perform a search for docker images and returns a table of
results.
}
