% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_stop}
\alias{docker_stop}
\title{Stop the active docker container}
\usage{
docker_stop(x)
}
\arguments{
\item{x}{either the name of the docker container or a docker_connection}
}
\description{
Stop the active docker container
}
