% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-helpers.R
\name{.sys_which}
\alias{.sys_which}
\title{Internal wrapper for Sys.which()}
\usage{
.sys_which(tool)
}
\description{
.sys_which adds Sys.which() inside containr's namespace so that
local_mock_bindings() can be used
}
\keyword{internal}
