% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{install_additional}
\alias{install_additional}
\title{Install Script}
\description{
This is a dataset containing stuff.
}
\section{install_additional.sh}{


This data is used in write_install_bash_file().
}

