% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{print.docker_connection}
\alias{print.docker_connection}
\title{Print method}
\usage{
\method{print}{docker_connection}(x, ...)
}
\arguments{
\item{x}{Name of image}

\item{...}{Additional params}
}
\description{
Print method
}
