% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{remove}
\alias{remove}
\title{Uninstall and remove a program}
\usage{
remove(id)
}
\arguments{
\item{id}{Program ID}
}
\value{
Logical
}
\description{
Uninstalls a program and removes it from your docker
}
\details{
If program is successfully removed from your Docker, TRUE is
returned else FALSE.
}
\seealso{
Other user: \code{\link{available}}, \code{\link{details}}
}
