% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{rocker_pull}
\alias{rocker_pull}
\title{Pull Rocker Images}
\usage{
rocker_pull(name, tag = NULL)
}
\arguments{
\item{name}{Rocker image name from the \link{data_rocker_table}}

\item{tag}{Version tag of rocker image. Will default to \code{latest}}
}
\description{
Pull Rocker Images
}
