% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R, R/run_rocker.R
\name{rocker_run}
\alias{rocker_run}
\title{RStudio Addin: Run active project in Rocker Rstudio docker container}
\usage{
rocker_run(
  image = "testing",
  tag = NULL,
  DISABLE_AUTH = TRUE,
  use_local = TRUE
)

rocker_run(
  image = "testing",
  tag = NULL,
  DISABLE_AUTH = TRUE,
  use_local = TRUE
)
}
\arguments{
\item{image}{Repository Name of rocker image from \link{data_rocker_table}. Defaults to
the \code{rstudio} stack \code{rocker/rstudio:latest}. For more information about each stack
visit \url{https://rocker-project.org/images/}. If you ran \code{docker_build()} then supply
this argument with the string of the \emph{name} of the build. Built container images can be
view by running the \code{docker_images()} function.}

\item{tag}{Version number of stack. Defaults to the \verb{:latest} tag if none is supplied.}

\item{DISABLE_AUTH}{Bypass authentication and show the R session. Defaults to TRUE and will
login when the session is launced in a browser.}

\item{use_local}{Whether to clone the local \code{.config}, \code{.Renviron} and \code{.Rprofile} to container session or
use the default settings. Defaults to \code{TRUE} and makes these config files available. Handy if the user has a
particular panel and theme layout and want to access private repos for install and testing.}
}
\description{
RStudio addin for Rocker containers

Launch the active R project into a Rocker Rstudio container
and opens a browser to the session.
}
\keyword{internal}
