% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containr-package.R
\docType{package}
\name{containr-package}
\alias{containr}
\alias{containr-package}
\title{containr: Containerize Your 'R' Project}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for containerizing 'R' projects. Reads an 'renv' lock file, generates a ready-to-use 'Dockerfile', builds the container image, and pushes it to a registry. Generated 'COPY' instructions preserve the local directory structure inside the container. Cross-platform builds are supported via 'Docker' buildx for targeting 'linux/amd64' from ARM hosts. Designed to help researchers build portable, reproducible workflows that can be reliably shared, archived, and rerun across systems. See R Core Team (2025) \url{https://www.R-project.org/}, Ushey et al. (2025) \url{https://CRAN.R-project.org/package=renv}, and Docker Inc. (2025) \url{https://www.docker.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erwinlares/containr}
  \item \url{https://erwinlares.github.io/containr/}
  \item Report bugs at \url{https://github.com/erwinlares/containr/issues}
}

}
\author{
\strong{Maintainer}: Erwin Lares \email{erwin.lares@wisc.edu} (\href{https://orcid.org/0000-0002-3284-828X}{ORCID})

Authors:
\itemize{
  \item Erwin Lares \email{erwin.lares@wisc.edu} (\href{https://orcid.org/0000-0002-3284-828X}{ORCID})
}

}
\keyword{internal}
