% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containr-package.R
\docType{package}
\name{containr-package}
\alias{containr}
\alias{containr-package}
\title{containr: Containerize Your 'R' Project}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automates the process of containerizing 'R' projects. The core function of 'containr' is 'generate_dockerfile()', which analyzes an 'R' project's environment and dependencies via an 'renv' lock file and generates a ready-to-use 'Dockerfile' that encapsulates the computational setup. The package helps researchers build portable and consistent workflows so that analyses can be reliably shared, archived, and rerun across systems. See R Core Team (2025) \url{https://www.R-project.org/}, Ushey et al. (2025) \url{https://CRAN.R-project.org/package=renv}, and Docker Inc. (2025) \url{https://www.docker.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erwinlares/containr}
  \item \url{https://erwinlares.github.io/containr/}
  \item Report bugs at \url{https://github.com/erwinlares/containr/issues}
}

}
\author{
\strong{Maintainer}: Erwin Lares \email{erwin.lares@wisc.edu}

}
\keyword{internal}
