% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containr-package.R
\docType{package}
\name{containr-package}
\alias{containr}
\alias{containr-package}
\title{containr: Containerize your R Project}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automate the process of containerizing R projects. Containr's core function, generate_dockerfile(), analyzes an R project’s environment and dependencies—via renv::renv.lock—and generates a ready-to-use Dockerfile that encapsulates the computational setup. The package is designed to assist researchers in building portable and consistent workflows, ensuring that analyses can be reliably shared, archived, and rerun across systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erwinlares/containr}
  \item \url{https://erwinlares.github.io/containr/}
  \item Report bugs at \url{https://github.com/erwinlares/containr/issues}
}

}
\author{
\strong{Maintainer}: Erwin Lares \email{erwin.lares@wisc.edu}

}
\keyword{internal}
