% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containr.R
\docType{package}
\name{containr}
\alias{containr}
\alias{containr-package}
\title{containr}
\details{
These functions allow a user to set the filepath for where the GenBank files
should be stored and verify the filepaths.
}
