% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containr.R
\name{containr}
\alias{containr}
\title{R6 Class for a ContainR}
\description{
A class object of \code{containr} that sets up a object to launch a defined rocker
containr in a browser.
}
\section{Rockered}{

Creates a new \code{containr} object based on the \url{https://rocker-project.org/images/}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{image}}{Name of the image from either the \link{data_rocker_table} or in docker register}

\item{\code{name}}{Name of the containr when launched. Defaults to the active Rstudio Project.}

\item{\code{tag}}{Tag of image defaults to \code{latest} if nothing set.
Start a ContainR}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-containr-new}{\code{containr$new()}}
\item \href{#method-containr-set_image}{\code{containr$set_image()}}
\item \href{#method-containr-set_name}{\code{containr$set_name()}}
\item \href{#method-containr-set_tag}{\code{containr$set_tag()}}
\item \href{#method-containr-set_login}{\code{containr$set_login()}}
\item \href{#method-containr-set_local}{\code{containr$set_local()}}
\item \href{#method-containr-start}{\code{containr$start()}}
\item \href{#method-containr-stop}{\code{containr$stop()}}
\item \href{#method-containr-status}{\code{containr$status()}}
\item \href{#method-containr-launch}{\code{containr$launch()}}
\item \href{#method-containr-proc}{\code{containr$proc()}}
\item \href{#method-containr-info}{\code{containr$info()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-new"></a>}}
\if{latex}{\out{\hypertarget{method-containr-new}{}}}
\subsection{Method \code{new()}}{
Setup and start a new \code{containr} in the background.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$new(
  image = NULL,
  name = NULL,
  tag = NULL,
  DISABLE_AUTH = TRUE,
  use_local = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{Repository Name of rocker image from \link{data_rocker_table}. Defaults to
the \code{rstudio} stack \code{rocker/rstudio:latest}. For more information about each stack
visit \url{https://rocker-project.org/images/}. If you ran \code{docker_build()} then supply
this argument with the string of the \emph{name} of the build. Built container images can be
view by running the \code{docker_images()} function.}

\item{\code{name}}{The name of the containr for the docker process. If none is applied then
it will default to the active Rstudio project name.}

\item{\code{tag}}{A character string of the require version. If no tag is supplied then the function will default to \code{latest}.}

\item{\code{DISABLE_AUTH}}{Bypass authentication and show the R session.
Defaults to TRUE and will login when the \code{launch()} function is invoked.}

\item{\code{use_local}}{If you want to port your local Rstudio session into the containr.
Defaults to \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-set_image"></a>}}
\if{latex}{\out{\hypertarget{method-containr-set_image}{}}}
\subsection{Method \code{set_image()}}{
Set or change the containr image name. Also checks the docker register for existing images.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$set_image(image)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{Set image name that was set from the \code{build_image()} setup with the \code{set_image_name}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-containr-set_name}{}}}
\subsection{Method \code{set_name()}}{
Change the containr name when launched. Defaults to active Rstudio project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Change containr name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-set_tag"></a>}}
\if{latex}{\out{\hypertarget{method-containr-set_tag}{}}}
\subsection{Method \code{set_tag()}}{
Change the containr name when launched. Defaults to active Rstudio project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$set_tag(tag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{Change tag name. Defaults to \code{latest}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-set_login"></a>}}
\if{latex}{\out{\hypertarget{method-containr-set_login}{}}}
\subsection{Method \code{set_login()}}{
Rstudio login
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$set_login(DISABLE_AUTH)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DISABLE_AUTH}}{Set to enable the Rstudio Login. Set to \code{TRUE} until secure
env is setup.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-set_local"></a>}}
\if{latex}{\out{\hypertarget{method-containr-set_local}{}}}
\subsection{Method \code{set_local()}}{
Port local config and theme settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$set_local(use_local)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_local}}{Flag \code{TRUE} to port local settings or \code{FALSE} for a clean session.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-start"></a>}}
\if{latex}{\out{\hypertarget{method-containr-start}{}}}
\subsection{Method \code{start()}}{
Start the predefined containr as set by the \code{image} argument.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-stop"></a>}}
\if{latex}{\out{\hypertarget{method-containr-stop}{}}}
\subsection{Method \code{stop()}}{
Stop containr object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-status"></a>}}
\if{latex}{\out{\hypertarget{method-containr-status}{}}}
\subsection{Method \code{status()}}{
Get the status of the containr
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-launch"></a>}}
\if{latex}{\out{\hypertarget{method-containr-launch}{}}}
\subsection{Method \code{launch()}}{
Opens the active containr in a browser.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$launch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-proc"></a>}}
\if{latex}{\out{\hypertarget{method-containr-proc}{}}}
\subsection{Method \code{proc()}}{
Displays the process information of the active containr
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$proc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-containr-info"></a>}}
\if{latex}{\out{\hypertarget{method-containr-info}{}}}
\subsection{Method \code{info()}}{
\code{info(containr)} or \code{containr$info()} shows some information about the
process on the screen, whether it is running and it's process id, etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{containr$info()}\if{html}{\out{</div>}}
}

}
}
