% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker}
\alias{docker}
\title{Docker Process Class}
\description{
Work in progress.

Docker System Wrapper WPI
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-docker-new}{\code{docker$new()}}
\item \href{#method-docker-print}{\code{docker$print()}}
\item \href{#method-docker-show_output}{\code{docker$show_output()}}
\item \href{#method-docker-docker_images}{\code{docker$docker_images()}}
\item \href{#method-docker-docker_containers}{\code{docker$docker_containers()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docker-new"></a>}}
\if{latex}{\out{\hypertarget{method-docker-new}{}}}
\subsection{Method \code{new()}}{
Work in progress: Eventually will be part of the containr class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$new(process = NA, commands = NA, options = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process}}{The system process. Plan is to add \code{podman} support but currently only
is setup for \code{docker}.}

\item{\code{commands}}{The main docker commands. Currently support commands
are \code{images}, \code{container} and \code{search}.}

\item{\code{options}}{the \code{command} option arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docker-print"></a>}}
\if{latex}{\out{\hypertarget{method-docker-print}{}}}
\subsection{Method \code{print()}}{
Prints the full command to the console for viewing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docker-show_output"></a>}}
\if{latex}{\out{\hypertarget{method-docker-show_output}{}}}
\subsection{Method \code{show_output()}}{
Displays returned output in \code{tibble} format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$show_output()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docker-docker_images"></a>}}
\if{latex}{\out{\hypertarget{method-docker-docker_images}{}}}
\subsection{Method \code{docker_images()}}{
Displays a \code{tibble} of local images.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$docker_images()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docker-docker_containers"></a>}}
\if{latex}{\out{\hypertarget{method-docker-docker_containers}{}}}
\subsection{Method \code{docker_containers()}}{
Displays a \code{tibble} of running containers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$docker_containers()}\if{html}{\out{</div>}}
}

}
}
