% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_push}
\alias{docker_push}
\title{Docker Push Image}
\usage{
docker_push(name, tag)
}
\arguments{
\item{name}{The user/name of the docker image}

\item{tag}{Version or tag of image. Defaults to latest.}
}
\description{
Push container image to \url{https://hub.docker.com/}.
}
