% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{docker_run}
\alias{docker_run}
\title{Run a Docker Container}
\usage{
docker_run(cmd, proc_args, name)
}
\arguments{
\item{cmd}{System command. \code{docker} as default until other container as are supported.}

\item{proc_args}{A character string of arguments to the \code{cmd}.}

\item{name}{Character string of the docker image name.}
}
\description{
Currently only \code{docker} is available but plans are to implement \code{podman} in the future.
}
