% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R, R/run_docker.R
\name{docker_run}
\alias{docker_run}
\title{RStudio Addin: Containerize and Render R Markdown Document with Docker}
\usage{
docker_run(rocker_name, tag = NULL)

docker_run(rocker_name, tag = NULL)
}
\arguments{
\item{rocker_name}{Name of rocker image from \link{data_rocker_table}}

\item{tag}{version number or use local R}
}
\description{
RStudio addin for containerizing and rendering the current document.

Launches current active R project into a rocker container
and opens a browser.
}
\keyword{internal}
