% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_search}
\alias{docker_search}
\title{Docker Search}
\usage{
docker_search(search = "rstudio")
}
\arguments{
\item{search}{A character string of image to search on DockerHub. Returns it in \code{tibble}
format.}
}
\description{
Docker Search
}
