% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dockerfile.R
\name{generate_dockerfile}
\alias{generate_dockerfile}
\title{Generate a Dockerfile for your R project}
\usage{
generate_dockerfile(
  verbose = FALSE,
  r_version = "latest",
  data_file = NULL,
  code_file = NULL,
  home_dir = "/home",
  install_quarto = FALSE,
  expose_port = "8787",
  r_mode = "base",
  comments = FALSE
)
}
\arguments{
\item{verbose}{logical (TRUE or FALSE). Should generate_dockerfile() print out progress? By default, it will silently create a Dockerfile}

\item{r_version}{a character string indicated a version of R, i.e., "4.3.0". By default, it will grab the latest version of R available}

\item{data_file}{a character string indication an optional name of a data file to copied into the container}

\item{code_file}{a character string indication an optional name of a script file to copied into the container}

\item{home_dir}{a character string specifying the home directory inside the container}

\item{install_quarto}{logical (TRUE or FALSE). If TRUE it will include supporting packages and system libraries to support Quarto and RMarkdown.}

\item{expose_port}{a character string indicating in which port will RStudio Server be accessible. It defaults to 8787}

\item{r_mode}{a character string. Inspired by the images in the Rocker Project. The options are "base" for base R, tidyverse, rstudio for RStudio Server, tidystudio which is tidyverse plus TeX Live and some publishing-related R packages}

\item{comments}{a logical (TRUE or FALSE). If TRUE, the Dockerfile generated will include comments detailing what each line does. If FALSE, the Dockerfile will be bare with only commands.}
}
\value{
invisibly returns NULL. This function is called for its side effects and does not return a value.
}
\description{
Generate a Dockerfile for your R project
}
\examples{
# Basic Usage
\dontrun{generate_dockerfile()}
# Specify a tidyverse ready image
\dontrun{generate_dockerfile(r_mode = "tidyverse")}
# Specify an image with R 4.2.0 installed
\dontrun{generate_dockerfile(r_version = "4.2.0")}
}
