% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R, R/run_docker.R
\name{rocker_run}
\alias{rocker_run}
\title{RStudio Addin: Run active project in Rocker Rstudio docker container}
\usage{
rocker_run(image = "rstudio", tag = NULL)

rocker_run(image = "rstudio", tag = NULL)
}
\arguments{
\item{image}{Repository Name of rocker image from \link{data_rocker_table}. Defaults to
the \code{rstudio} stack \code{rocker/rstudio:latest}. For more information about each stack
visit \url{https://rocker-project.org/images/}. If you ran \code{docker_build()} then supply
this argument with the string of the \emph{name} of the build. Built container images can be
view by running the \code{docker_list()} function.}

\item{tag}{Version number of stack. Defaults to the \verb{:latest} tag if none is supplied.}
}
\description{
RStudio addin for Rocker containers

Launch the active R project into a Rocker Rstudio container
and opens a browser to the session. Mounts local config and .Renviron from local
session.
}
\keyword{internal}
